Address PR review: data-driven AI_MODELS, composer-owned dialog, utils split
- AI_MODELS is now an array of `{id, label, family}` so `getAiModelFamily`
  is a lookup instead of a string-prefix heuristic. New `getAiModelLabel`
  helper replaces direct `AI_MODELS[id]` access at all call sites; new
  models become a one-line data change instead of two.
- The cross-family confirmation dialog moves from `SessionView` into the
  composer (where the model picker lives) — composer owns the orchestration
  and renders the dialog inline; `FamilySwitchConfirmDialog` extracted to
  its own file for testability.
- Single-caller helpers in `apps/cli/ai/tools/utils.ts` move to where
  they're used: `splitCommandArgs` → `wp-cli.ts` (private), and
  `formatInvalidBlocks` → `validate-blocks.ts` (private). The four
  preview tools share `runPreviewCommand` via a new `preview-helpers.ts`
  rather than the general utils grab-bag.
