+++ apps/cli/ai/tools/utils.ts
@@ -1,82 +1,7 @@
 import path from 'path';
-import { type ValidationReport } from 'cli/ai/block-validator';
 import { readCliConfig, type SiteData } from 'cli/lib/cli-config/core';
 import { getSiteByFolder } from 'cli/lib/cli-config/sites';
 import { getProgressCallback, setProgressCallback } from 'cli/logger';
-
-/**
- * Splits a command string into arguments, respecting quoted strings.
- * Handles both single and double quotes, e.g.:
- *   post create --post_title="Ember & Oak" --post_type=page
- *   → ['post', 'create', '--post_title=Ember & Oak', '--post_type=page']
- */
-function splitBasicCommandArgs( command: string ): string[] {
-	const args: string[] = [];
-	let current = '';
-	let inQuote: string | null = null;
-
-	for ( let i = 0; i < command.length; i++ ) {
-		const char = command[ i ];
-
-		if ( inQuote ) {
-			if ( char === inQuote ) {
-				inQuote = null;
-			} else {
-				current += char;
-			}
-		} else if (
-			( char === '"' || char === "'" ) &&
-			( current === '' || current.endsWith( '=' ) )
-		) {
-			inQuote = char;
-		} else if ( /\s/.test( char ) ) {
-			if ( current ) {
-				args.push( current );
-				current = '';
-			}
-		} else {
-			current += char;
-		}
-	}
-
-	if ( current ) {
-		args.push( current );
-	}
-
-	return args;
-}
-
-function stripMatchingOuterQuotes( value: string ): string {
-	if ( value.length < 2 ) {
-		return value;
-	}
-
-	const firstChar = value[ 0 ];
-	const lastChar = value[ value.length - 1 ];
-	if ( ( firstChar === '"' || firstChar === "'" ) && firstChar === lastChar ) {
-		return value.slice( 1, -1 );
-	}
-
-	return value;
-}
-
-export function splitCommandArgs( command: string ): string[] {
-	const postContentMarker = '--post_content=';
-	const postContentIndex = command.indexOf( postContentMarker );
-
-	// Large block content is commonly emitted without shell quoting and should
-	// be treated as a single literal argument that consumes the rest of the command.
-	if ( postContentIndex !== -1 ) {
-		const prefix = command.slice( 0, postContentIndex ).trim();
-		const postContent = stripMatchingOuterQuotes(
-			command.slice( postContentIndex + postContentMarker.length ).trim()
-		);
-
-		return [ ...splitBasicCommandArgs( prefix ), `${ postContentMarker }${ postContent }` ];
-	}
-
-	return splitBasicCommandArgs( command );
-}
 
 async function findSiteByName( name: string ): Promise< SiteData | undefined > {
 	const config = await readCliConfig();
@@ -113,23 +38,6 @@ export function textResult( text: string ) {
 	return {
 		content: [ { type: 'text' as const, text } ],
 	};
-}
-
-export function formatInvalidBlocks( report: ValidationReport ): string[] {
-	const lines: string[] = [];
-	for ( const result of report.results ) {
-		if ( ! result.isValid ) {
-			lines.push( `  - ${ result.blockName }` );
-			for ( const issue of result.issues ) {
-				lines.push( `    ${ issue }` );
-			}
-			if ( result.expectedContent !== undefined ) {
-				lines.push( `    Expected: ${ result.expectedContent }` );
-				lines.push( `    Actual:   ${ result.originalContent }` );
-			}
-		}
-	}
-	return lines;
 }
 
 /**
@@ -202,25 +110,4 @@ export async function captureCommandOutput( fn: () => Promise< void > ): Promise
 		progressOutput: progressMessages.join( '\n' ),
 		exitCode,
 	};
-}
-
-export async function runPreviewCommand(
-	fn: () => Promise< void >,
-	fallbackMessage: string,
-	errorPrefix: string
-) {
-	try {
-		const result = await captureCommandOutput( fn );
-		const output = result.consoleOutput || result.progressOutput || fallbackMessage;
-
-		if ( result.exitCode ) {
-			return errorResult( output );
-		}
-
-		return textResult( output );
-	} catch ( error ) {
-		return errorResult(
-			`${ errorPrefix }: ${ error instanceof Error ? error.message : String( error ) }`
-		);
-	}
 }
