+++ apps/ui/src/components/session-view/composer/index.tsx
@@ -1,13 +1,18 @@
-import { AI_MODELS } from '@studio/common/ai/models';
+import { AI_MODELS, getAiModelFamily, getAiModelLabel } from '@studio/common/ai/models';
 import { AI_SKILL_COMMANDS } from '@studio/common/ai/slash-commands';
+import { useQueryClient } from '@tanstack/react-query';
+import { useNavigate } from '@tanstack/react-router';
 import { __ } from '@wordpress/i18n';
 import { arrowUp, chevronDownSmall } from '@wordpress/icons';
 import { Icon } from '@wordpress/ui';
 import { useCallback, useState } from 'react';
 import * as Menu from '@/components/menu';
+import { useConnector } from '@/data/core';
+import { SESSIONS_QUERY_KEY } from '@/data/queries/use-sessions';
 import { EnvironmentPill } from './environment-pill';
+import { FamilySwitchConfirmDialog } from './family-switch-confirm-dialog';
 import styles from './style.module.css';
-import type { AiModelId, SyncSite } from '@/data/core';
+import type { AiModelId, AiSessionEvent, LoadedAiSession, SyncSite } from '@/data/core';
 
 /**
  * Invisible structural placeholder that mirrors Composer's outer DOM (shell +
@@ -35,7 +40,6 @@ interface ComposerProps {
 	isInterrupting?: boolean;
 	error: string | null;
 	model: AiModelId;
-	onModelChange: ( model: AiModelId ) => void;
 	onSend: ( prompt: string ) => Promise< void >;
 	onInterrupt: () => Promise< void >;
 	// Environment pill: only rendered when both a `sessionId` and a linked
@@ -44,6 +48,15 @@ interface ComposerProps {
 	sessionId?: string;
 	effectiveEnvironment?: 'local' | 'live';
 	liveSite?: SyncSite;
+	// Session events drive the cross-family confirmation logic — we skip the
+	// dialog when the session is empty (nothing to lose) and use them for the
+	// optimistic `session.model_selected` cache update on same-family swaps.
+	events?: AiSessionEvent[];
+	// Local owner site id, when the session is anchored to one. Required to
+	// spin up a fresh session via `connector.createSession` on a confirmed
+	// family swap; if absent we fall back to the in-place model change so the
+	// dropdown still works for unowned sessions.
+	ownerSiteId?: string;
 }
 
 const isMacPlatform =
@@ -54,14 +67,24 @@ export function Composer( {
 	isInterrupting = false,
 	error,
 	model,
-	onModelChange,
 	onSend,
 	onInterrupt,
 	sessionId,
 	effectiveEnvironment = 'local',
 	liveSite,
+	events,
+	ownerSiteId,
 }: ComposerProps ) {
 	const [ value, setValue ] = useState( '' );
+	const connector = useConnector();
+	const queryClient = useQueryClient();
+	const navigate = useNavigate();
+
+	// Cross-family swap state. We hold the picked model here while the
+	// confirmation dialog is open; nothing is persisted until the user
+	// confirms.
+	const [ pendingFamilyChange, setPendingFamilyChange ] = useState< AiModelId | null >( null );
+	const [ familySwitchInFlight, setFamilySwitchInFlight ] = useState( false );
 
 	const send = useCallback( async () => {
 		const trimmed = value.trim();
@@ -80,6 +103,93 @@ export function Composer( {
 		}
 	}, [ value, onSend ] );
 
+	// Same-family swap: optimistically append a `session.model_selected` event
+	// so the composer reflects the new pick immediately. The main process
+	// writes the same event to the JSONL; if that write fails we fall back
+	// to the prior state via query invalidation.
+	const applySameFamilyModel = useCallback(
+		( picked: AiModelId ) => {
+			if ( ! sessionId ) {
+				return;
+			}
+			const timestamp = new Date().toISOString();
+			queryClient.setQueryData< LoadedAiSession >(
+				[ ...SESSIONS_QUERY_KEY, sessionId ],
+				( prev ) =>
+					prev
+						? {
+								...prev,
+								events: [
+									...prev.events,
+									{ type: 'session.model_selected', timestamp, model: picked },
+								],
+						  }
+						: prev
+			);
+			void connector.setSessionModel( sessionId, picked ).catch( () => {
+				void queryClient.invalidateQueries( {
+					queryKey: [ ...SESSIONS_QUERY_KEY, sessionId ],
+				} );
+			} );
+		},
+		[ connector, queryClient, sessionId ]
+	);
+
+	const handleModelChange = useCallback(
+		( picked: AiModelId ) => {
+			if ( picked === model ) {
+				return;
+			}
+			// Cross-family switch: defer until the user confirms in the dialog
+			// — the runtimes don't share a transcript, so continuing the same
+			// JSONL across families would make the on-screen history disagree
+			// with the agent's actual memory. We skip the prompt when:
+			//   - the session has no user turns yet (nothing to lose), or
+			//   - the session has no local owner site (we have no `siteId` to
+			//     pass to `createSession`, so the fresh-session path can't
+			//     run; fall back to the in-place switch instead of blocking
+			//     the dropdown).
+			const hasTurns = ( events ?? [] ).some( ( event ) => event.type === 'user.message' );
+			if ( getAiModelFamily( model ) !== getAiModelFamily( picked ) && ownerSiteId && hasTurns ) {
+				setPendingFamilyChange( picked );
+				return;
+			}
+			applySameFamilyModel( picked );
+		},
+		[ applySameFamilyModel, events, model, ownerSiteId ]
+	);
+
+	const cancelFamilyChange = useCallback( () => {
+		if ( familySwitchInFlight ) {
+			return;
+		}
+		setPendingFamilyChange( null );
+	}, [ familySwitchInFlight ] );
+
+	const confirmFamilyChange = useCallback( async () => {
+		if ( ! pendingFamilyChange || ! ownerSiteId ) {
+			return;
+		}
+		setFamilySwitchInFlight( true );
+		try {
+			const newSession = await connector.createSession( ownerSiteId );
+			// Persist the model on the fresh session before navigating so the
+			// composer there opens already on the picked family —
+			// `setSessionModel` writes a `session.model_selected` event the
+			// new view picks up via `resolveSessionModel`. If this fails we
+			// still navigate; the user can re-pick from the new view's
+			// dropdown.
+			await connector
+				.setSessionModel( newSession.id, pendingFamilyChange )
+				.catch( () => undefined );
+			void queryClient.invalidateQueries( { queryKey: SESSIONS_QUERY_KEY } );
+			setPendingFamilyChange( null );
+			void navigate( { to: '/sessions/$sessionId', params: { sessionId: newSession.id } } );
+		} finally {
+			setFamilySwitchInFlight( false );
+		}
+	}, [ connector, navigate, ownerSiteId, pendingFamilyChange, queryClient ] );
+
 	const canSend = value.trim().length > 0;
 	const placeholder = busy
 		? __( 'Queue a follow-up instruction…' )
@@ -88,122 +198,129 @@ export function Composer( {
 	const modKey = isMacPlatform ? '⌘' : 'Ctrl';
 
 	return (
-		<div className={ styles.root }>
-			<div className={ styles.shell }>
-				<textarea
-					className={ styles.input }
-					placeholder={ placeholder }
-					value={ value }
-					onChange={ ( event ) => setValue( event.target.value ) }
-					onKeyDown={ ( event ) => {
-						if ( event.key === 'Enter' && ( event.metaKey || event.ctrlKey ) ) {
-							event.preventDefault();
-							void send();
-						}
-					} }
-					rows={ 2 }
-				/>
-				<div className={ styles.toolbar }>
-					<div className={ styles.leftActions }>
-						<Menu.Root modal={ false }>
-							<Menu.Trigger
-								render={
-									<button
-										type="button"
-										className={ `${ styles.iconButton } ${ styles.glyphButton }` }
-										aria-label={ __( 'Commands' ) }
+		<>
+			<div className={ styles.root }>
+				<div className={ styles.shell }>
+					<textarea
+						className={ styles.input }
+						placeholder={ placeholder }
+						value={ value }
+						onChange={ ( event ) => setValue( event.target.value ) }
+						onKeyDown={ ( event ) => {
+							if ( event.key === 'Enter' && ( event.metaKey || event.ctrlKey ) ) {
+								event.preventDefault();
+								void send();
+							}
+						} }
+						rows={ 2 }
+					/>
+					<div className={ styles.toolbar }>
+						<div className={ styles.leftActions }>
+							<Menu.Root modal={ false }>
+								<Menu.Trigger
+									render={
+										<button
+											type="button"
+											className={ `${ styles.iconButton } ${ styles.glyphButton }` }
+											aria-label={ __( 'Commands' ) }
+										>
+											/
+										</button>
+									}
+								/>
+								<Menu.Popup side="top" align="start" className={ styles.commandsMenuPopup }>
+									{ AI_SKILL_COMMANDS.map( ( command ) => (
+										<Menu.Item
+											key={ command.name }
+											onClick={ () => {
+												void onSend( `/${ command.name }` );
+											} }
+										>
+											<span className={ styles.commandItem }>
+												<span className={ styles.commandName }>/{ command.name }</span>
+												<span className={ styles.commandDescription }>{ command.description }</span>
+											</span>
+										</Menu.Item>
+									) ) }
+								</Menu.Popup>
+							</Menu.Root>
+						</div>
+						<div className={ styles.rightActions }>
+							{ sessionId && liveSite ? (
+								<EnvironmentPill
+									sessionId={ sessionId }
+									effectiveEnvironment={ effectiveEnvironment }
+									liveSite={ liveSite }
+									disabled={ busy }
+								/>
+							) : null }
+							<Menu.Root modal={ false }>
+								<Menu.Trigger
+									render={
+										<button
+											type="button"
+											className={ styles.pill }
+											aria-label={ __( 'Select model' ) }
+										>
+											<span>{ getAiModelLabel( model ) }</span>
+											<Icon icon={ chevronDownSmall } size={ 16 } />
+										</button>
+									}
+								/>
+								<Menu.Popup side="top" align="end">
+									<Menu.RadioGroup
+										value={ model }
+										onValueChange={ ( value ) => handleModelChange( value as AiModelId ) }
 									>
-										/
-									</button>
-								}
-							/>
-							<Menu.Popup side="top" align="start" className={ styles.commandsMenuPopup }>
-								{ AI_SKILL_COMMANDS.map( ( command ) => (
-									<Menu.Item
-										key={ command.name }
-										onClick={ () => {
-											void onSend( `/${ command.name }` );
-										} }
-									>
-										<span className={ styles.commandItem }>
-											<span className={ styles.commandName }>/{ command.name }</span>
-											<span className={ styles.commandDescription }>{ command.description }</span>
-										</span>
-									</Menu.Item>
-								) ) }
-							</Menu.Popup>
-						</Menu.Root>
-					</div>
-					<div className={ styles.rightActions }>
-						{ sessionId && liveSite ? (
-							<EnvironmentPill
-								sessionId={ sessionId }
-								effectiveEnvironment={ effectiveEnvironment }
-								liveSite={ liveSite }
-								disabled={ busy }
-							/>
-						) : null }
-						<Menu.Root modal={ false }>
-							<Menu.Trigger
-								render={
-									<button
-										type="button"
-										className={ styles.pill }
-										aria-label={ __( 'Select model' ) }
-									>
-										<span>{ AI_MODELS[ model ] }</span>
-										<Icon icon={ chevronDownSmall } size={ 16 } />
-									</button>
-								}
-							/>
-							<Menu.Popup side="top" align="end">
-								<Menu.RadioGroup
-									value={ model }
-									onValueChange={ ( value ) => onModelChange( value as AiModelId ) }
-								>
-									{ ( Object.entries( AI_MODELS ) as [ AiModelId, string ][] ).map(
-										( [ id, label ] ) => (
+										{ AI_MODELS.map( ( { id, label } ) => (
 											<Menu.RadioItem key={ id } value={ id }>
 												{ label }
 											</Menu.RadioItem>
-										)
-									) }
-								</Menu.RadioGroup>
-							</Menu.Popup>
-						</Menu.Root>
-						{ busy ? (
+										) ) }
+									</Menu.RadioGroup>
+								</Menu.Popup>
+							</Menu.Root>
+							{ busy ? (
+								<button
+									type="button"
+									className={ styles.stopButton }
+									onClick={ () => void onInterrupt() }
+									aria-label={ isInterrupting ? __( 'Stopping' ) : __( 'Stop' ) }
+									aria-busy={ isInterrupting }
+									title={
+										isInterrupting ? __( 'Stopping… click again to force stop' ) : __( 'Stop' )
+									}
+								>
+									<span className={ styles.stopGlyph } aria-hidden="true" />
+								</button>
+							) : null }
 							<button
 								type="button"
-								className={ styles.stopButton }
-								onClick={ () => void onInterrupt() }
-								aria-label={ isInterrupting ? __( 'Stopping' ) : __( 'Stop' ) }
-								aria-busy={ isInterrupting }
-								title={
-									isInterrupting ? __( 'Stopping… click again to force stop' ) : __( 'Stop' )
-								}
+								className={ styles.sendButton }
+								onClick={ () => void send() }
+								disabled={ ! canSend }
+								aria-label={ sendAriaLabel }
 							>
-								<span className={ styles.stopGlyph } aria-hidden="true" />
+								<Icon icon={ arrowUp } size={ 18 } />
 							</button>
-						) : null }
-						<button
-							type="button"
-							className={ styles.sendButton }
-							onClick={ () => void send() }
-							disabled={ ! canSend }
-							aria-label={ sendAriaLabel }
-						>
-							<Icon icon={ arrowUp } size={ 18 } />
-						</button>
+						</div>
 					</div>
 				</div>
+				<div className={ styles.meta }>
+					<span className={ styles.metaHint }>
+						{ modKey }↩ { __( 'to send' ) } · shift↩ { __( 'for newline' ) }
+					</span>
+					{ error ? <span className={ styles.error }>{ error }</span> : null }
+					<span className={ styles.metaUses }>{ __( 'Uses 1 message' ) }</span>
+				</div>
 			</div>
-			<div className={ styles.meta }>
-				<span className={ styles.metaHint }>
-					{ modKey }↩ { __( 'to send' ) } · shift↩ { __( 'for newline' ) }
-				</span>
-				{ error ? <span className={ styles.error }>{ error }</span> : null }
-				<span className={ styles.metaUses }>{ __( 'Uses 1 message' ) }</span>
-			</div>
-		</div>
+			<FamilySwitchConfirmDialog
+				currentModel={ model }
+				pendingModel={ pendingFamilyChange }
+				inFlight={ familySwitchInFlight }
+				onCancel={ cancelFamilyChange }
+				onConfirm={ () => void confirmFamilyChange() }
+			/>
+		</>
 	);
 }
