+++ apps/ui/src/components/session-view/index.tsx
@@ -1,18 +1,8 @@
-import { AI_MODELS, getAiModelFamily, resolveSessionModel } from '@studio/common/ai/models';
-import { useQueryClient } from '@tanstack/react-query';
-import { useNavigate } from '@tanstack/react-router';
-import { __, sprintf } from '@wordpress/i18n';
-import { Button, Dialog, IconButton } from '@wordpress/ui';
+import { resolveSessionModel } from '@studio/common/ai/models';
+import { __ } from '@wordpress/i18n';
+import { IconButton } from '@wordpress/ui';
 import { clsx } from 'clsx';
-import {
-	useCallback,
-	useLayoutEffect,
-	useMemo,
-	useRef,
-	useState,
-	type ReactNode,
-	type Ref,
-} from 'react';
+import { useLayoutEffect, useMemo, useRef, useState, type ReactNode, type Ref } from 'react';
 import { Composer, ComposerSkeleton } from '@/components/session-view/composer';
 import { pickLiveSite } from '@/components/session-view/composer/environment-pill';
 import { Conversation } from '@/components/session-view/conversation';
@@ -20,20 +10,15 @@ import { EmptyBackground } from '@/components/session-view/empty-background';
 import { QueuedPrompts } from '@/components/session-view/queued-prompts';
 import { SiteDropdown } from '@/components/site-dropdown';
 import { SitePreview } from '@/components/site-preview';
-import { useConnector } from '@/data/core';
 import { useAgentRun } from '@/data/queries/use-agent-run';
 import { useConnectedWpcomSites } from '@/data/queries/use-connected-wpcom-sites';
-import {
-	SESSIONS_QUERY_KEY,
-	useSession,
-	useSessionEffectiveEnvironment,
-} from '@/data/queries/use-sessions';
+import { useSession, useSessionEffectiveEnvironment } from '@/data/queries/use-sessions';
 import { useSites } from '@/data/queries/use-sites';
 import { useFullscreen } from '@/hooks/use-fullscreen';
 import { useSidebarCollapsed } from '@/hooks/use-sidebar-collapsed';
 import { drawerIcon } from '@/lib/icons';
 import styles from './style.module.css';
-import type { AiModelId, AiSessionSummary, LoadedAiSession } from '@/data/core';
+import type { AiSessionSummary } from '@/data/core';
 
 interface SessionHeaderProps {
 	summary: AiSessionSummary;
@@ -121,9 +106,6 @@ function SessionFrame( { header, composer, preview, scrollRef, children }: Sessi
 
 export function SessionView( { sessionId }: { sessionId: string } ) {
 	const { data, isLoading, error } = useSession( sessionId );
-	const connector = useConnector();
-	const queryClient = useQueryClient();
-	const navigate = useNavigate();
 	const { data: sites } = useSites();
 	const ownerSitePath = data?.summary.ownerSitePath;
 	const ownerSite = ownerSitePath
@@ -147,95 +129,6 @@ export function SessionView( { sessionId }: { sessionId: string } ) {
 		removeQueuedPrompt,
 	} = useAgentRun( sessionId );
 	const currentModel = useMemo( () => resolveSessionModel( data?.events ?? [] ), [ data?.events ] );
-	// Cross-family switches need a fresh session: the runtimes don't share a
-	// transcript, so continuing the same JSONL with a different family makes
-	// the on-screen history disagree with the agent's actual memory. We hold
-	// the user's pick here and surface a confirmation dialog before acting.
-	const [ pendingFamilyChange, setPendingFamilyChange ] = useState< AiModelId | null >( null );
-	const [ familySwitchInFlight, setFamilySwitchInFlight ] = useState( false );
-
-	// Same-family switch: optimistically append a `session.model_selected` event
-	// so the composer reflects the new pick immediately. The main process writes
-	// the same event to the JSONL; if that write fails we fall back to the
-	// prior state.
-	const applySameFamilyModel = useCallback(
-		( model: AiModelId ) => {
-			const timestamp = new Date().toISOString();
-			queryClient.setQueryData< LoadedAiSession >(
-				[ ...SESSIONS_QUERY_KEY, sessionId ],
-				( prev ) =>
-					prev
-						? {
-								...prev,
-								events: [ ...prev.events, { type: 'session.model_selected', timestamp, model } ],
-						  }
-						: prev
-			);
-			void connector.setSessionModel( sessionId, model ).catch( () => {
-				void queryClient.invalidateQueries( {
-					queryKey: [ ...SESSIONS_QUERY_KEY, sessionId ],
-				} );
-			} );
-		},
-		[ connector, queryClient, sessionId ]
-	);
-
-	const onModelChange = useCallback(
-		( model: AiModelId ) => {
-			if ( model === currentModel ) {
-				return;
-			}
-			// Cross-family switch: defer until the user confirms in the dialog —
-			// the runtimes don't share a transcript, so continuing the same JSONL
-			// across families would make the on-screen history disagree with the
-			// agent's actual memory. We skip the prompt when:
-			//   - the session has no user turns yet (nothing to lose), or
-			//   - the session has no local owner site (we have no `siteId` to pass
-			//     to `createSession`, so the fresh-session path can't run; fall
-			//     back to the in-place switch instead of blocking the dropdown).
-			const hasTurns = ( data?.events ?? [] ).some( ( event ) => event.type === 'user.message' );
-			if (
-				getAiModelFamily( currentModel ) !== getAiModelFamily( model ) &&
-				ownerSite &&
-				hasTurns
-			) {
-				setPendingFamilyChange( model );
-				return;
-			}
-			applySameFamilyModel( model );
-		},
-		[ applySameFamilyModel, currentModel, data?.events, ownerSite ]
-	);
-
-	const cancelFamilyChange = useCallback( () => {
-		if ( familySwitchInFlight ) {
-			return;
-		}
-		setPendingFamilyChange( null );
-	}, [ familySwitchInFlight ] );
-
-	const confirmFamilyChange = useCallback( async () => {
-		if ( ! pendingFamilyChange || ! ownerSite ) {
-			return;
-		}
-		setFamilySwitchInFlight( true );
-		try {
-			const newSession = await connector.createSession( ownerSite.id );
-			// Persist the model on the fresh session before navigating so the
-			// composer there opens already on the picked family — `setSessionModel`
-			// writes a `session.model_selected` event the new view picks up via
-			// `resolveSessionModel`. If this fails we still navigate; the user
-			// can re-pick from the new view's dropdown.
-			await connector
-				.setSessionModel( newSession.id, pendingFamilyChange )
-				.catch( () => undefined );
-			void queryClient.invalidateQueries( { queryKey: SESSIONS_QUERY_KEY } );
-			setPendingFamilyChange( null );
-			void navigate( { to: '/sessions/$sessionId', params: { sessionId: newSession.id } } );
-		} finally {
-			setFamilySwitchInFlight( false );
-		}
-	}, [ connector, navigate, ownerSite, pendingFamilyChange, queryClient ] );
 	const pendingQuestionTexts = useMemo(
 		() => new Set( pendingQuestions.map( ( q ) => q.question ) ),
 		[ pendingQuestions ]
@@ -291,116 +184,49 @@ export function SessionView( { sessionId }: { sessionId: string } ) {
 	}
 
 	return (
-		<>
-			<SessionFrame
-				scrollRef={ scrollRef }
-				header={
-					<SessionHeader
-						summary={ data.summary }
-						previewOpen={ showPreview }
-						onTogglePreview={ () => setPreviewOpen( ( open ) => ! open ) }
-						canTogglePreview={ canTogglePreview }
-					/>
-				}
-				composer={
-					<div className={ styles.column }>
-						<QueuedPrompts prompts={ queuedPrompts } onRemove={ removeQueuedPrompt } />
-						<Composer
-							busy={ composerBusy }
-							isInterrupting={ isInterrupting }
-							error={ runError }
-							model={ currentModel }
-							onModelChange={ onModelChange }
-							onSend={ sendMessage }
-							onInterrupt={ interrupt }
-							sessionId={ sessionId }
-							effectiveEnvironment={ effectiveEnvironment }
-							liveSite={ liveSite }
-						/>
-					</div>
-				}
-				preview={
-					showPreview && ownerSite ? (
-						<SitePreview site={ ownerSite } sessionId={ sessionId } />
-					) : null
-				}
-			>
-				{ isEmpty ? <EmptyBackground /> : null }
-				<div className={ clsx( styles.column, styles.conversationSpacing ) }>
-					<Conversation
-						data={ data }
-						isRunning={ isRunning }
-						startedAt={ startedAt }
-						pendingQuestions={ pendingQuestionTexts }
-						pendingAnswers={ pendingAnswers }
-						onAnswerQuestion={ answerQuestion }
+		<SessionFrame
+			scrollRef={ scrollRef }
+			header={
+				<SessionHeader
+					summary={ data.summary }
+					previewOpen={ showPreview }
+					onTogglePreview={ () => setPreviewOpen( ( open ) => ! open ) }
+					canTogglePreview={ canTogglePreview }
+				/>
+			}
+			composer={
+				<div className={ styles.column }>
+					<QueuedPrompts prompts={ queuedPrompts } onRemove={ removeQueuedPrompt } />
+					<Composer
+						busy={ composerBusy }
+						isInterrupting={ isInterrupting }
+						error={ runError }
+						model={ currentModel }
+						onSend={ sendMessage }
+						onInterrupt={ interrupt }
+						sessionId={ sessionId }
+						effectiveEnvironment={ effectiveEnvironment }
+						liveSite={ liveSite }
+						events={ data.events }
+						ownerSiteId={ ownerSite?.id }
 					/>
 				</div>
-			</SessionFrame>
-			<FamilySwitchConfirmDialog
-				currentModel={ currentModel }
-				pendingModel={ pendingFamilyChange }
-				inFlight={ familySwitchInFlight }
-				onCancel={ cancelFamilyChange }
-				onConfirm={ () => void confirmFamilyChange() }
-			/>
-		</>
-	);
-}
-
-function FamilySwitchConfirmDialog( {
-	currentModel,
-	pendingModel,
-	inFlight,
-	onCancel,
-	onConfirm,
-}: {
-	currentModel: AiModelId;
-	pendingModel: AiModelId | null;
-	inFlight: boolean;
-	onCancel: () => void;
-	onConfirm: () => void;
-} ) {
-	return (
-		<Dialog.Root
-			open={ pendingModel !== null }
-			onOpenChange={ ( next ) => {
-				if ( ! next ) {
-					onCancel();
-				}
-			} }
+			}
+			preview={
+				showPreview && ownerSite ? <SitePreview site={ ownerSite } sessionId={ sessionId } /> : null
+			}
 		>
-			<Dialog.Popup size="small">
-				<Dialog.Header>
-					<Dialog.Title>{ __( 'Start a new conversation?' ) }</Dialog.Title>
-				</Dialog.Header>
-				<p>
-					{ pendingModel
-						? sprintf(
-								/* translators: 1: current model name, 2: new model name */
-								__(
-									'Switching from %1$s to %2$s starts a fresh conversation — the two model families don\u2019t share memory. Your current chat stays in the sidebar.'
-								),
-								AI_MODELS[ currentModel ],
-								AI_MODELS[ pendingModel ]
-						  )
-						: '' }
-				</p>
-				<Dialog.Footer>
-					<Dialog.Action variant="minimal" tone="neutral" disabled={ inFlight }>
-						{ __( 'Cancel' ) }
-					</Dialog.Action>
-					<Button
-						variant="solid"
-						tone="brand"
-						loading={ inFlight }
-						loadingAnnouncement={ __( 'Starting new conversation' ) }
-						onClick={ onConfirm }
-					>
-						{ __( 'Start new conversation' ) }
-					</Button>
-				</Dialog.Footer>
-			</Dialog.Popup>
-		</Dialog.Root>
+			{ isEmpty ? <EmptyBackground /> : null }
+			<div className={ clsx( styles.column, styles.conversationSpacing ) }>
+				<Conversation
+					data={ data }
+					isRunning={ isRunning }
+					startedAt={ startedAt }
+					pendingQuestions={ pendingQuestionTexts }
+					pendingAnswers={ pendingAnswers }
+					onAnswerQuestion={ answerQuestion }
+				/>
+			</div>
+		</SessionFrame>
 	);
 }
